Implementado comportamento para sincronizar notas nao sincronizadas: Implementada query no Dao; Salvando nao sincronizadas no banco de dados e logo apos chamando o metodo de sincronizacao; Atualizando o metodo de atualizar na activity para sincronizar.

+++ app/src/main/java/br/com/alura/ceep/repository/NotaRepository.kt
@@ -30,7 +30,10 @@ class NotaRepository(
     }
 
     suspend fun remove(id: String) {
-        dao.remove(id)
+        dao.desativa(id)
+        if (webClient.remove(id)) {
+            dao.remove(id)
+        }
     }
 
     suspend fun salva(nota: Nota) {
@@ -42,6 +45,10 @@ class NotaRepository(
     }
 
     suspend fun sincroniza() {
+        val notasDesativadas = dao.buscaDesativadas().first()
+        notasDesativadas.forEach { notaDesativada ->
+            remove(notaDesativada.id)
+        }
         val notasNaoSincronizadas = dao.buscaNaoSincronizadas().first()
         notasNaoSincronizadas.forEach { notaNaoSincronizada ->
             salva(notaNaoSincronizada)
